Extract bump_version code
