+++ xtask/src/main.rs
@@ -11,7 +11,6 @@ use esp_metadata::{Chip, Config};
 use strum::IntoEnumIterator;
 use xtask::{
     Package,
-    Version,
     cargo::{CargoAction, CargoArgsBuilder},
     commands::*,
 };
@@ -41,26 +40,6 @@ enum Cli {
     Publish(PublishArgs),
     /// Generate git tags for all new package releases.
     TagReleases(TagReleasesArgs),
-}
-
-#[derive(Debug, Args)]
-struct BumpVersionArgs {
-    /// How much to bump the version by.
-    ///
-    /// If the version is a pre-release, this will just remove the pre-release
-    /// version tag.
-    #[arg(value_enum)]
-    amount: Version,
-    /// Pre-release version to append to the version.
-    ///
-    /// If the package is already a pre-release, this will ignore `amount`. If
-    /// the package is not a pre-release, this will bump the version by
-    /// `amount` and append the pre-release version as `<pre>.0`.
-    #[arg(long)]
-    pre: Option<String>,
-    /// Package(s) to target.
-    #[arg(value_enum, default_values_t = Package::iter())]
-    packages: Vec<Package>,
 }
 
 #[derive(Debug, Args)]
@@ -163,15 +142,6 @@ fn main() -> Result<()> {
 
 // ----------------------------------------------------------------------------
 // Subcommands
-
-fn bump_version(workspace: &Path, args: BumpVersionArgs) -> Result<()> {
-    // Bump the version by the specified amount for each given package:
-    for package in args.packages {
-        xtask::bump_version(workspace, package, args.amount, args.pre.as_deref())?;
-    }
-
-    Ok(())
-}
 
 fn fmt_packages(workspace: &Path, args: FmtPackagesArgs) -> Result<()> {
     let mut packages = args.packages;
